Fix PLD combos broken by 6.3 rework
This makes the folowing changes:
1. Removes the Goring Blade Combo and Goring Blade Atonement Feature
   options.<br>
   *(Goring blade is no longer part of any combos and these no longer work
   with paladin's new rotation.)*
2. Fix the Requiescat Confiteor option.
3. Make the Confiteor Feature option unconditionally switch to the Confiteor
   combo when under the effect of Requiescat.<br>
   *(With the changes, it never makes sense to use Requiescat for Holy Spirit
   or Holy Circle.)*

The Requiescat Confiteor and Confiteor Feature options are effectively
identical now, they just apply to different base actions. The implementations
could be combined fairly easily, but I was uneasy about making that kind of
structural change. It might also make sense to remove the
`[SecretCustomCombo]` attribute from the Confiteor Feature option (and maybe
rename it as well), but I have not done so here.